requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
certifi==2025.4.26
charset-normalizer==3.4.1
customtkinter==5.2.2
darkdetect==0.8.0
diskcache==5.6.3
exceptiongroup==1.2.2
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.24
langchain-core==0.3.56
langchain-text-splitters==0.3.8
langsmith==0.3.39
llama_cpp_python==0.3.8
MarkupSafe==3.0.2
numpy==2.2.5
orjson==3.10.18
packaging==24.2
pydantic==2.11.4
pydantic_core==2.33.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.40
tenacity==9.1.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
zstandard==0.23.0